Removing old agent code
