+++ src/api.ts
@@ -23,10 +23,10 @@ export const play = (request: Request): Response => {
     switch (request.type) {
         case RequestType.USER_INPUT:
             if (!request.content) {
-                throw new RegalError(ErrorCode.INVALID_INPUT, "Request content must be supplied with a user input request.");
+                throw new RegalError("Request content must be supplied with a user input request.");
             }
             if (!request.game) {
-                throw new RegalError(ErrorCode.INVALID_INPUT, "A game instance must be supplied with a user input request.");
+                throw new RegalError("A game instance must be supplied with a user input request.");
             }
 
             request.game.output = [];
@@ -35,14 +35,14 @@ export const play = (request: Request): Response => {
             break;
         
         case RequestType.ALARM:
-            throw new RegalError(ErrorCode.NOT_YET_IMPLEMENTED, "Alarms not yet supported.");
+            throw new RegalError("Alarms not yet supported.");
 
         case RequestType.START:
              game = Game.onGameStart();
              break;
 
         default:
-            throw new RegalError(ErrorCode.INVALID_INPUT, "Invalid request type.");
+            throw new RegalError("Invalid request type.");
     }
 
     return { game, output: game.output };
