+++ src/index.ts
@@ -4,4 +4,3 @@ import * as Event from './event';
 export { Api };
 export { Event };
 export * from './game';
-export * from './agent';
